+++ src/main/java/lotto/controller/LottoController.java
@@ -1,7 +1,7 @@
 package lotto.controller;
 
 import java.util.List;
-import lotto.domain.Rank;
+import lotto.domain.lotto.utils.Rank;
 import lotto.domain.lotto.LottoLine;
 import lotto.domain.lotto.LottoNumber;
 import lotto.domain.lotto.LottoResult;
