+++ src/test/java/domain/RankTest.java
@@ -2,7 +2,7 @@ package domain;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-import lotto.domain.Rank;
+import lotto.domain.lotto.utils.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
