+++ src/test/java/domain/lotto/LottoLineTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 import java.util.Arrays;
 import java.util.List;
-import lotto.domain.Rank;
+import lotto.domain.lotto.utils.Rank;
 import lotto.domain.lotto.LottoLine;
 import lotto.domain.lotto.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
